requirements.txt
pytorch-transformers
nltk
torch
pandas
numpy
matplotlib
tweet-preprocessor